removed support for dashed names when using Element.prototype.setStyle() or Element.prototype.styles(), simply make sure you pass javascript compliant style properties.
Changed test for seeing if style can be applied form
try{
this.style[prop] = val;
} catch(e){}

to
if(typeof this.style[prop] == 'string'){
	this.style[prop] = val;
}

which is now more efficient when using it rapidly e.g. for effects

Also added suppport for styleFloat.  Simply set cssFloat with Element.prototype.setStyle() or Element.prototype.styles() and it will be converted to styleFloat for non-standard IE