*added support for PCRE based routing patterns to all controllers, this is controlled by setting the $routing_patterns property of any controller that extends sb_Controller which is all of them by default
e.g. 

class IndexController extends sb_Controller{

	public $routing_patterns = Array(
		'~/\d:\d+~' => 'route_digits'
	);
	
	public function route_digits($pattern=''){
		print_r(Gateway::$request);
	}

}

this would pass requests for http://yoursite.com/1:23 to the route_digits method of the IndexController

You can set it to pass to any callable, the default for string is to look for a function that matches, if none exists then look for a local method.

You can even set these dynamically in the __construct()

e.g.
public function __construct(){
	$this->patterns['~/\d:\d+~' ] = function($pattern){//do something};
}